Add plugin init hook Move plugin to own module
