+++ pytui/runner.py
@@ -14,6 +14,8 @@ import unittest
 from StringIO import StringIO
 import __builtin__
 
+from plugin import PytestPlugin
+
 logger = logging.getLogger(__name__)
 
 
@@ -139,40 +141,6 @@ class TestResultWindow(urwid.LineBox):
 
     def set_focus(self, item):
         self._original_widget.set_focus(item)
-
-class PytestPlugin(object):
-    def __init__(self, runner, test_list={}):
-        self.runner = runner
-
-    def pytest_runtest_protocol(self, item, nextitem):
-        logger.debug('pytest_runtest_protocol %s %s', item, nextitem)
-
-    def pytest_itemcollected(self, item):
-        logger.debug('pytest_itemcollected %s', item)
-        self.runner.add_test(item)
-
-    def pytest_collectstart(self, collector):
-        logger.debug('pytest_collectstart %s', collector)
-
-
-    def pytest_runtest_makereport(self, item, call):
-        logger.debug('pytest_runtest_makereport %s %s', item, str(type(call.excinfo)))
-        if call.excinfo:
-            self.runner.set_exc_info(item.nodeid, call.excinfo)
-
-    def pytest_runtest_logreport(self, report):
-        if (report.when == 'setup' and report.outcome == 'skipped'
-            or report.when == 'call'):
-            self.runner.set_test_result(report.nodeid, report, report.capstdout + report.capstderr)
-
-    def pytest_collectreport(self, report):
-        logger.debug('pytest_collectreport %s', report)
-
-    def pytest_collection_modifyitems(self, session, config, items):
-        logger.debug('pytest_collection_modifyitems %s %s %s', session, config, items)
-
-        items[:] = filter(self.runner.is_test_filtered, items)
-        logger.debug('Filtered items %s', items)
 
 
 class Runner(object):
@@ -555,6 +523,7 @@ class TestRunnerUI(object):
         elif key == 'meta up':
             self.focus_failed_sibling(-1)
 
+
 def main():
     path = sys.argv[1] if len(sys.argv) - 1 else '.'
     logging_tools.configure()
@@ -563,10 +532,6 @@ def main():
     runner = PytestRunner(path)
     ui = TestRunnerUI(runner)
     ui.run()
-
-
-runner = PytestRunner('.')
-plugin = PytestPlugin(runner)
 
 
 if __name__ == '__main__':
